requirements.txt
fastapi~=0.99.1
uvicorn~=0.22.0
pydantic~=1.10.10
pyCryptodome~=3.10.1
bcrypt~=4.0.1
tortoise-orm~=0.19.3
aioredis~=2.0.1
aiofiles~=23.1.0
aiohttp~=3.8.4
PyJWT~=2.7.0
python-multipart~=0.0.6
aiomysql~=0.2.0
